+++ lib/yt/models/snippet.rb
@@ -159,14 +159,6 @@ module Yt
         @complete ||= data.fetch :complete, true
       end
 
-      # @return [Boolean] if the resource is a video category, whether the
-      #   category is assignable based on the given filters.
-      # @return [nil] if the resource is a channel.
-      # @return [nil] if the resource is a playlist.
-      # @return [nil] if the resource is a video.
-      # @return [nil] if the resource is a playlist item.
-      has_attribute :assignable
-
     private
 
       has_attribute :thumbnails, default: {}