+++ spec/requests/as_account/video_spec.rb
@@ -64,6 +64,7 @@ describe Yt::Video, :device_app do
       expect(video.scheduled_end_time).to be_nil
       expect(video.concurrent_viewers).to be_nil
       expect(video.embed_html).to be_a String
+      expect(video.category_title).to be_a String
     end
 
     it { expect{video.update}.to fail }