+++ CHANGELOG.md
@@ -10,6 +10,7 @@ For more information about changelogs, check
 ## 0.15.3 - unreleased
 
 * [FEATURE] New `file_size`, `file_type`, `container` methods for Yt::Video.
+* [BUGFIX] Retrieve `category_id` also for videos obtained through a search.
 
 ## 0.15.2 - 2015-04-27